% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linReg.R
\name{getPred.lm.boston}
\alias{getPred.lm.boston}
\title{Construir un JSON}
\usage{
getPred.lm.boston(lstat, age)
}
\arguments{
\item{lstat}{....}

\item{df}{A dataframe.
Devuelve el resultado de una regresion lineal simple ajustada al dataset "BOSTON" del paquete MASS ()}

\item{agre}{....}
}
\description{
Construir un JSON
}
