% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{getPred.lm.s}
\alias{getPred.lm.s}
\title{Devuelve el resultado de una regresion lineal simple ajustada al dataset "BOSTON" del paquete MASS ()}
\usage{
getPred.lm.s(lstat)
}
\arguments{
\item{df}{A dataframe.}
}
\description{
Devuelve el resultado de una regresion lineal simple ajustada al dataset "BOSTON" del paquete MASS ()
}
