% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{getPred.ranger.pima}
\alias{getPred.ranger.pima}
\title{TidyModels RandomForest Indian Pima Diabetes}
\usage{
getPred.ranger.pima(
  pregnant,
  glucose,
  pressure,
  triceps,
  insulin,
  mass,
  pedigree,
  age
)
}
\arguments{
\item{pregnant}{Numeric, number of times pregnant}

\item{glucose}{Numeric, plasma glucose concentration a 2 hours in an oral glucose tolerance test}

\item{pressure}{Numeric, Diastolic blood pressure (mm Hg)}

\item{triceps}{Numeric, Triceps skin fold thickness (mm)}

\item{insulin}{Numeric, 2-Hour serum insulin (mu U/ml)}

\item{mass}{Numeric, body mass index (weight in kg/(height in m)\^2)}

\item{pedigree}{Numeric, diabetes pedigree function}

\item{age}{Numeric, age (yers)}
}
\value{
Logical: True si diabetico, Falso si no.
}
\description{
Pronosticar si una indigena pima es diabetica con un random Forest ajustado al dataset IndianPimaDiabetes del paquete MLBench.
}
\examples{
getPred.ranger.pima(pregnant = 2, glucose = 95, pressure = 70, triceps = 31, insulin = 102, mass = 28.2, pedigree = 0.67, age = 23)
}
\seealso{
\code{\link[parsnip]{rand_forest}}

\code{\link{getPred.lm.boston}}
}
\author{
Yago
\email{josevicente.yago@um.es}
}
