% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadModel}
\alias{loadModel}
\title{loadModel: Cargar modelos preetrenados}
\usage{
loadModel(model.name)
}
\arguments{
\item{model.name}{String}
}
\description{
Los modelos se cargan del directorio /inst/extdata para hacer las predicciones
}
