% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyML.R
\docType{package}
\name{dummyML}
\alias{dummyML}
\title{dummyML: Un paquete de juguete para demostrar como desplegar ML en R.}
\description{
El paquete dummyML ofrece una serie de modelos preentrenados, listos para ser utilizados.
}
\section{dummyML functions}{

The foo functions getPred.lm.s, getPred.lm.m
}

